Breaking up s3 module in multiple files.
